ArchiveSite and RandomSite are now independent components
